+++ MochiKit/Format.js
@@ -29,6 +29,100 @@ MochiKit.Format.toString = function () {
     return this.__repr__();
 };
 
+MochiKit.Format.numberFormatter = function (pattern, placeholder/* = "" */, locale/* = "default" */) {
+    // http://java.sun.com/docs/books/tutorial/i18n/format/numberpattern.html
+    // | 0 | leading or trailing zeros
+    // | # | just the number
+    // | , | separator
+    // | . | decimal separator
+    // | % | Multiply by 100 and format as percent
+    if (typeof(placeholder) == "undefined") {
+        placeholder = "";
+    }
+    var match = pattern.match(/((?:[0#]+,)?[0#]+)(?:\.([0#]+))?(%)?/);
+    if (!match) {
+        throw TypeError("Invalid pattern");
+    }
+    var header = pattern.substr(0, match.index);
+    var footer = pattern.substr(match.index + match[0].length);
+    if (header.search(/-/) == -1) {
+        header = header + "-";
+    }
+    var whole = match[1];
+    var frac = (typeof(match[2]) == "string") ? match[2] : "";
+    var isPercent = (typeof(match[3]) == "string");
+    var tmp = whole.split(/,/);
+    var separatorAt;
+    if (typeof(locale) == "undefined") {
+        locale = "default";
+    }
+    if (tmp.length == 1) {
+        separatorAt = null;
+    } else {
+        separatorAt = tmp[1].length;
+    }
+    var leadingZeros = whole.length - whole.replace(/0+/, "").length;
+    var trailingZeros = frac.length - frac.replace(/0+/, "").length;
+    var precision = frac.length;
+    return function (num) {
+        if (typeof(num) == "undefined" || num == null || isNaN(num)) {
+            return placeholder;
+        }
+        var curheader = header;
+        var curfooter = footer;
+        if (num < 0) {
+            num = -num;
+        } else {
+            curheader = curheader.replace(/-/, "");
+        }
+        var me = arguments.callee;
+        var fmt = MochiKit.Format.formatLocale(locale);
+        if (isPercent) {
+            num = num * 100.0;
+            curfooter = fmt.percent + curfooter;
+        }
+        num = MochiKit.Format.roundToFixed(num, precision);
+        var parts = num.split(/\./);
+        var whole = parts[0];
+        var frac = (parts.length == 1) ? "" : parts[1];
+        var res = "";
+        while (whole.length < leadingZeros) {
+            whole = "0" + whole;
+        }
+        if (separatorAt) {
+            while (whole.length > separatorAt) {
+                var i = whole.length - separatorAt;
+                res = res + fmt.separator + whole.substring(i, whole.length);
+                whole = whole.substring(0, i);
+            }
+        }
+        res = whole + res;
+        if (precision > 0) {
+            while (frac.length < trailingZeros) {
+                frac = frac + "0";
+            }
+            res = res + fmt.decimal + frac;
+        }
+        return curheader + res + curfooter;
+    };
+};
+
+MochiKit.Format.formatLocale = function (locale) {
+    if (typeof(locale) == "undefined" || locale == null) {
+        locale = "default";
+    }
+    if (typeof(locale) == "string") {
+        var rval = MochiKit.Format.LOCALE[locale];
+        if (typeof(rval) == "string") {
+            rval = arguments.callee(rval);
+            MochiKit.Format.LOCALE[locale] = rval;
+        }
+        return rval;
+    } else {
+        return locale;
+    }
+};
+
 MochiKit.Format.twoDigitAverage = function (numerator, denominator) {
     /***
 
@@ -124,6 +218,8 @@ MochiKit.Format.percentFormat = function (someFloat) {
 MochiKit.Format.EXPORT = [
     "truncToFixed",
     "roundToFixed",
+    "numberFormatter",
+    "formatLocale",
     "twoDigitAverage",
     "twoDigitFloat",
     "percentFormat",
@@ -131,6 +227,13 @@ MochiKit.Format.EXPORT = [
     "rstrip",
     "strip"
 ];
+
+MochiKit.Format.LOCALE = {
+    "en_US": {"separator": ",", "decimal": ".", "percent": "%"},
+    "de_DE": {"separator": ".", "decimal": ",", "percent": "%"},
+    "fr_FR": {"separator": " ", "decimal": ",", "percent": "%"},
+    "default": "en_US"
+};
 
 MochiKit.Format.EXPORT_OK = [];
 MochiKit.Format.EXPORT_TAGS = {